refactor: SetupStepをViewModel内に定義するよう変更
